+++ hybrid_vision/hybrid_vision_android/lib/hybrid_vision_android.dart
@@ -1,11 +1,2 @@
-import 'package:hybrid_vision_platform_interface/hybrid_vision_platform_interface.dart';
-
-import 'src/barcode.dart';
-import 'src/jni.dart';
-
-abstract class HybridVisionAndroid {
-  static void registerWith() {
-    Jni.initDLApi();
-    BarcodePlatform.instance = BarcodePlatformImpl();
-  }
-}
+export 'src/hybrid_vision_android_plugin.dart';
+export 'src/barcode_detection.dart';