+++ hybrid_vision/hybrid_vision_platform_interface/lib/hybrid_vision_platform_interface.dart
@@ -1,2 +1,2 @@
 export 'src/core.dart';
-export 'src/barcode.dart';
+export 'src/barcode_detection.dart';