+++ hybrid_vision/hybrid_vision_android/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.1.0
+
+* Move `HybridVisionAndroid` to `HybridVisionAndroidPlugin`.
+* Move `BarcodePlatformImpl` to `AndroidBarcodeDetectionPlatform`.
+* Move `BarcodeDetectorImpl` to `AndroidBarcodeDetector`.
+* Optimize project structure.
+
 ## 0.0.1
 
 * Implement `BarcodePlatform` interface.